✨ 홈 및 피드 화면 분리, 화면 이동 링크 추가
