+++ messages/en.json
@@ -1,5 +1,12 @@
 {
   "HomePage": {
-    "title": "Hello world!"
+    "title": "Ever.Log",
+    "description": "I aim to create products that resonate with everyone.\nTo build better products, I focus on seeing from the user’s perspective.",
+    "feed": "Feed",
+    "github": "https://github.com/owen-ever"
+  },
+  "PostListPage": {
+    "title": "Feed",
+    "home": "Home"
   }
 }
