+++ src/app/[locale]/posts/[slug]/page.tsx
