+++ src/app/layout.tsx
@@ -11,8 +11,8 @@ const notoSansKR = Noto_Sans_KR({
 });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Ever.Log',
+  description: 'Ever.Log',
 };
 
 export default async function RootLayout({
