+++ src/components/markdown/AppMarkdown.css
@@ -53,7 +53,7 @@
   /* 하이라이트 설정 */
   strong {
     font-weight: 700;
-    
+
     &.highlight {
       background-color: #bef9be;
     }
@@ -77,7 +77,8 @@
   }
 
   /* 밑줄 + 강조 조합 */
-  u strong, strong u {
+  u strong,
+  strong u {
     font-weight: 700;
     text-decoration: underline;
   }
@@ -99,7 +100,8 @@
   }
 
   /* 리스트 설정 */
-  ul, ol {
+  ul,
+  ol {
     margin: 1rem 0;
     padding-left: 2rem;
   }
@@ -203,7 +205,8 @@
     margin: 1rem 0;
   }
 
-  th, td {
+  th,
+  td {
     border: 1px solid #d1d5db;
     padding: 0.5rem;
     text-align: left;
