+++ src/lib/formatter.ts
@@ -1,0 +1,3 @@
+export function formatLine(text: string) {
+  return text.split('\n');
+}
